+++ src/GiftVoucher.php
@@ -2,6 +2,7 @@
 namespace verbb\giftvoucher;
 
 use verbb\giftvoucher\adjusters\GiftVoucherAdjuster;
+use verbb\giftvoucher\adjusters\GiftVoucherShippingAdjuster;
 use verbb\giftvoucher\assetbundles\GiftVoucherAsset;
 use verbb\giftvoucher\base\PluginTrait;
 use verbb\giftvoucher\elements\Code;
@@ -250,27 +251,27 @@ class GiftVoucher extends Plugin
     private function _registerAdjusters()
     {
         Event::on(OrderAdjustments::class, OrderAdjustments::EVENT_REGISTER_ORDER_ADJUSTERS, function(RegisterComponentTypesEvent $event) {
-            $settings = $this->getSettings();
+            // Because of how discount adjusters work, we can't modify the shipping value, which in almost
+            // all cases, we want to include. The regular discount adjuster will take care of the total
+            // item value just fine, but won't effect the shipping. We need to create another (duplicate)
+            // adjuster that provides a discount against the shipping amount, but as a shipping adjuster
+            // (not a discount adjuster). This then subtracts the shipping amount from the voucher.
+            // SO - we need to insert our shipping adjuster after regular discounts, before Tax.
+            $types = $event->types;
 
-            // Re-order the built-in adjusters to ensure gift vouchers are applied before tax.
-            if ($settings->registerAdjuster === 'beforeTax') {
-                $types = $event->types;
+            // Find the Tax adjuster, it should go before that, but if its not found (Commerce Lite), append
+            $taxKey = array_search(Tax::class, $event->types);
 
-                // Find the Tax adjuster, it should go before that, but if its not found (Commerce Lite), append
-                $taxKey = array_search(Tax::class, $event->types);
-
-                if ($taxKey) {
-                    array_splice($types, $taxKey, 0, GiftVoucherAdjuster::class);
-                } else {
-                    $types[] = GiftVoucherAdjuster::class;
-                }
-
-                $event->types = $types;
-            } elseif ($settings->registerAdjuster === 'afterTax') {
-                $event->types[] = GiftVoucherAdjuster::class;
+            if ($taxKey) {
+                array_splice($types, $taxKey, 0, GiftVoucherShippingAdjuster::class);
+            } else {
+                $types[] = GiftVoucherShippingAdjuster::class;
             }
+
+            $event->types = $types;
         });
 
+        // Apply the discount adjuster normally, where it should go in adjuster order.
         Event::on(OrderAdjustments::class, OrderAdjustments::EVENT_REGISTER_DISCOUNT_ADJUSTERS, function(RegisterComponentTypesEvent $event) {
             $event->types[] = GiftVoucherAdjuster::class;
         });